sickle_macros
Derive Macro EventHandler

#[derive(EventHandler)]

Macro to derive an EntityCommand to register and store one-shot systems.

Dependencies

[lib]
proc-macro = true

[dependencies]
syn = "1.0"
quote = "1.0"
proc-macro2 = "1.0"

See: https://doc.rust-lang.org/book/ch19-06-macros.html#procedural-macros-for-generating-code-from-attributes

Usage

See it in action: https://youtu.be/s1lQD-R_kqg

#[derive(Component, EventHandler)]
pub struct OnSomething {
  system_id: SystemId,
  pub active: bool,
}

impl Plugin for SomethingPlugin {
  fn build(&self, app: &mut App) {
      app.add_systems(Update, handle_pressed);
  }
}

// Call this system when you need to do something
fn handle_pressed(
    mut interaction_query: Query<(Entity, &Interaction), (Changed<Interaction>, With<OnSomething>)>,
    mut handler_query: Query<&mut OnSomething>,
    mut commands: Commands,
) {
    for (entity, interaction) in &mut interaction_query {
        let handler = handler_query.get_mut(entity).unwrap().into_inner();
        if *interaction == Interaction::Pressed {
            let mut active_handler = handler.clone();
            active_handler.active = true;
            commands.entity(entity).insert(active_handler);
            commands.run_system(handler.system_id);
            commands.entity(entity).insert(handler.clone());
        }
    }
}

// Spawn a UI button from e.g. a Startup system. `MainMenuRootNode` would be a marker component
// somewhere in the app, used to mark a container.
fn spawn_something_button(mut commands: Commands, root_node: Query<Entity, With<MainMenuRootNode>>) {
    commands.entity(root_node.single()).with_children(|parent| {
        parent
            .spawn(ButtonBundle {
                style: Style {
                    width: Val::Px(200.),
                    border: UiRect::all(Val::Px(2.)),
                    ..default()
                },
                background_color: Color::rgb(0.35, 0.35, 0.35).into(),
                ..default()
            })
            // A component can only be added once, but the handler can be any system
            //.add(OnSomethingHandler::from(||println!("Button pressed")))
            //.add(OnSomethingHandler::from(do_the_something_general))
            .add(OnSomethingHandler::from(do_the_something_with_the_button));
    });
}

fn do_the_something_general(){
    // Do something without caring which button is pressed
}

// Do something for the exact button being pressed
fn do_the_something_with_the_button(
  mut pressed_query: Query<(Entity, &OnSomething), With<OnSomething>>,
) {
  for (entity, on_something) in &mut pressed_query {
    if on_something.active {
      // Do something!
    }
  }
}